backend/pyproject.toml
[project]
name = "polymarket-monitor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "httpx==0.27.2",
  "websocket-client==1.7.0",
  "sqlmodel==0.0.22",
  "psycopg==3.2.3",
  "pydantic-settings==2.5.2",
  "apscheduler==3.10.4",
  "feedparser==6.0.11",
  "openai==1.51.2",
]

[project.optional-dependencies]
dev = ["pytest==8.3.3", "respx==0.21.1"]

[tool.pytest.ini_options]
testpaths = ["tests"]
